Change calendar2 to calendar, remove deprectated calendar component.
