+++ docs/app/documentation/component-docs/calendar/examples/calendar-i18n-example.component.ts
@@ -1,7 +1,7 @@
 import { CalendarI18n } from '../../../../../../library/src/lib/calendar/i18n/calendar-i18n';
 import { Component, Injectable } from '@angular/core';
 import { CalendarI18nLabels } from '../../../../../../library/src/lib/calendar/i18n/calendar-i18n-labels';
-import { FdDate } from '../../../../../../library/src/lib/calendar/calendar2/models/fd-date';
+import { FdDate } from '../../../../../../library/src/lib/calendar/models/fd-date';
 
 const localized_values = {
     'fr': {
@@ -53,7 +53,8 @@ export class CustomI18nLabels extends CalendarI18nLabels {
 
 @Component({
     selector: 'fd-calendar-i18n-example',
-    template: `<fd-calendar2 [(ngModel)]="selectedDay"></fd-calendar2>`,
+    template: `
+        <fd-calendar [(ngModel)]="selectedDay"></fd-calendar>`,
 
     // Note that this can be provided in the root of your application.
     providers: [
