+++ docs/app/documentation/component-docs/calendar/examples/calendar-monday-start-example.component.ts
@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
-import { FdDate } from '../../../../../../library/src/lib/calendar/calendar2/models/fd-date';
+import { FdDate } from '../../../../../../library/src/lib/calendar/models/fd-date';
 
 @Component({
     selector: 'fd-calendar-monday-start-example',
-    template: `<fd-calendar2 [calType]="'single'" [(ngModel)]="selectedDay" [startingDayOfWeek]="1"></fd-calendar2>`
+    template: `
+        <fd-calendar [calType]="'single'" [(ngModel)]="selectedDay" [startingDayOfWeek]="1"></fd-calendar>`
 })
 export class CalendarMondayStartExampleComponent {
     selectedDay = {
