+++ docs/app/documentation/component-docs/calendar/examples/calendar-single-example.component.ts
@@ -1,16 +1,16 @@
 import { Component } from '@angular/core';
-import { FdDate } from '../../../../../../library/src/lib/calendar/calendar2/models/fd-date';
+import { FdDate } from '../../../../../../library/src/lib/calendar/models/fd-date';
 import { FormControl, FormGroup } from '@angular/forms';
 
 @Component({
     selector: 'fd-calendar-single-example',
     template: `
-        <fd-calendar2 [calType]="'single'"
+        <fd-calendar [calType]="'single'"
                      [(ngModel)]="selectedDay"
                      [blockFunction]="myBlockFunction"
                      [disableFunction]="myDisableFunction">
-        </fd-calendar2>
-        
+        </fd-calendar>
+
         <button fd-button (click)="disableWednesday()">Disable Wednesday</button>
         <br/><br/>
         <div>Selected Date: {{selectedDay.date.toDateString()}}</div>`
