+++ docs/app/documentation/component-docs/date-picker/examples/date-picker-disabled-example.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { FdDate } from '../../../../../../library/src/lib/calendar/calendar2/models/fd-date';
+import { FdDate } from '../../../../../../library/src/lib/calendar/models/fd-date';
 
 @Component({
     selector: 'fd-date-picker-disabled-example',
