+++ docs/app/documentation/component-docs/date-picker/examples/date-picker-i18n-example.component.ts
@@ -1,7 +1,7 @@
 import { CalendarI18n } from '../../../../../../library/src/lib/calendar/i18n/calendar-i18n';
 import { Component, Injectable } from '@angular/core';
 import { CalendarI18nLabels } from '../../../../../../library/src/lib/calendar/i18n/calendar-i18n-labels';
-import { FdDate } from '../../../../../../library/src/lib/calendar/calendar2/models/fd-date';
+import { FdDate } from '../../../../../../library/src/lib/calendar/models/fd-date';
 
 const localized_values = {
     'fr': {
