+++ library/src/lib/calendar/calendar-views/calendar2-month-view/calendar2-month-view.component.html
