--- library/src/lib/calendar/calendar.component.scss
@@ -1,7 +1,0 @@
-:host {
-    display: block;
-}
-
-.fd-calendar__content {
-    background-color: white;
-}
