+++ library/src/lib/calendar/calendar.module.ts
@@ -2,19 +2,19 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IconModule } from '../icon/icon.module';
 
+import { CalendarHeaderComponent } from './calendar-header/calendar-header.component';
+import { CalendarDayViewComponent } from './calendar-views/calendar-day-view/calendar-day-view.component';
+import { Calendar2MonthViewComponent } from './calendar-views/calendar2-month-view/calendar2-month-view.component';
+import { Calendar2YearViewComponent } from './calendar-views/calendar2-year-view/calendar2-year-view.component';
+import { CalendarService } from './calendar.service';
 import { CalendarComponent } from './calendar.component';
-import { Calendar2Component } from './calendar2/calendar2.component';
-import { Calendar2HeaderComponent } from './calendar2/calendar2-header/calendar2-header.component';
-import { Calendar2DayViewComponent } from './calendar2/calendar2-views/calendar2-day-view/calendar2-day-view.component';
-import { Calendar2MonthViewComponent } from './calendar2/calendar2-views/calendar2-month-view/calendar2-month-view.component';
-import { Calendar2YearViewComponent } from './calendar2/calendar2-views/calendar2-year-view/calendar2-year-view.component';
-import { Calendar2Service } from './calendar2/calendar2.service';
 
 @NgModule({
-    declarations: [CalendarComponent, Calendar2Component, Calendar2HeaderComponent,
-        Calendar2DayViewComponent, Calendar2MonthViewComponent, Calendar2YearViewComponent],
+    declarations: [CalendarComponent, CalendarHeaderComponent,
+        CalendarDayViewComponent, Calendar2MonthViewComponent, Calendar2YearViewComponent],
     imports: [CommonModule, IconModule],
-    exports: [CalendarComponent, Calendar2Component],
-    providers: [Calendar2Service]
+    exports: [CalendarComponent, CalendarDayViewComponent,
+        CalendarHeaderComponent, Calendar2YearViewComponent, Calendar2MonthViewComponent],
+    providers: [CalendarService]
 })
 export class CalendarModule {}
