+++ library/src/lib/calendar/public_api.ts
@@ -1,4 +1,3 @@
 export * from './calendar.module';
 export * from './i18n/calendar-i18n';
-export * from './format/date-parser';
 export * from './i18n/calendar-i18n-labels';
