+++ library/src/lib/date-picker/public_api.ts
@@ -1,1 +1,2 @@
 export * from './date-picker.module';
+export * from './format/date-parser';
