+++ library/src/lib/datetime-picker/datetime-picker.component.html
@@ -30,19 +30,19 @@
             [attr.aria-hidden]="!isOpen"
             [style.display]="'block'">
             <div class="fd-datetime__container">
-                <fd-calendar2 calType="single"
-                              (closeCalendar)="closePopover()"
-                              [disableFunction]="disableFunction ? disableFunction : null"
-                              [blockFunction]="blockFunction ? blockFunction : null"
-                              [disableRangeStartFunction]="disableRangeStartFunction ? disableRangeStartFunction : null"
-                              [disableRangeEndFunction]="disableRangeEndFunction ? disableRangeEndFunction : null"
-                              [blockRangeStartFunction]="blockRangeStartFunction ? blockRangeStartFunction : null"
-                              [blockRangeEndFunction]="blockRangeEndFunction ? blockRangeEndFunction : null"
-                              [selectedDate]="selectedDate"
-                              (selectedDateChange)="handleDateChange($event)"
-                              (isValidDateChange)="isInvalidDateInputHandler($event)"
-                              [escapeFocusFunction]="null"
-                              [startingDayOfWeek]="startingDayOfWeek"></fd-calendar2>
+                <fd-calendar calType="single"
+                             (closeCalendar)="closePopover()"
+                             [disableFunction]="disableFunction ? disableFunction : null"
+                             [blockFunction]="blockFunction ? blockFunction : null"
+                             [disableRangeStartFunction]="disableRangeStartFunction ? disableRangeStartFunction : null"
+                             [disableRangeEndFunction]="disableRangeEndFunction ? disableRangeEndFunction : null"
+                             [blockRangeStartFunction]="blockRangeStartFunction ? blockRangeStartFunction : null"
+                             [blockRangeEndFunction]="blockRangeEndFunction ? blockRangeEndFunction : null"
+                             [selectedDate]="selectedDate"
+                             (selectedDateChange)="handleDateChange($event)"
+                             (isValidDateChange)="isInvalidDateInputHandler($event)"
+                             [escapeFocusFunction]="null"
+                             [startingDayOfWeek]="startingDayOfWeek"></fd-calendar>
                 <div class="fd-datetime__separator"></div>
                 <fd-time [disabled]="disabled"
                          [meridian]="meridian"
