+++ library/src/lib/datetime-picker/format/datetime-parser.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { FdDatetime } from '../models/fd-datetime';
-import { FdDate } from '../../calendar/calendar2/models/fd-date';
+import { FdDate } from '../../calendar/models/fd-date';
 import { TimeObject } from '../../time/time-object';
 
 export function DATE_TIME_FORMAT_FACTORY() {
@@ -17,14 +17,14 @@ export function DATE_TIME_FORMAT_FACTORY() {
 export abstract class DateTimeFormatParser {
 
     /**
-     * Should take in a string value and return a date object.
-     * @param value String to concert to a date object.
+     * Should take in a string value and return a FdDatetime model object.
+     * @param value String to concert to a FdDatetime model object.
      */
     abstract parse(value: string): FdDatetime;
 
     /**
-     * Should take in a date object and return a string representation.
-     * @param date String to concert to a date object.
+     * Should take in a FdDatetime model object and return a string representation.
+     * @param date FdDatetime object to concert to a date string.
      */
     abstract format(date: FdDatetime): string;
 }
@@ -36,8 +36,8 @@ export abstract class DateTimeFormatParser {
 export class DateTimeFormatParserDefault extends DateTimeFormatParser {
 
     /**
-     * Takes in a string representation of a date and returns a Date object.
-     * @param value String to convert to a date.
+     * Takes in a string representation of a date and returns a FdDatetime object.
+     * @param value String to convert to a FdDatetime model object.
      */
     public parse(value: string): FdDatetime {
         if (!value) {
@@ -62,8 +62,8 @@ export class DateTimeFormatParserDefault extends DateTimeFormatParser {
     }
 
     /**
-     * Takes in a date object and returns the string representation.
-     * @param date Date object to convert to a string.
+     * Takes in a FdDatetime object and returns the string representation.
+     * @param date FdDatetime model object to convert to a string.
      */
     public format(date: FdDatetime): string {
         return date.day + '.' +
