+++ library/src/lib/datetime-picker/models/fd-datetime.ts
@@ -1,16 +1,30 @@
-import { FdDate } from '../../calendar/calendar2/models/fd-date';
+/**
+ * Default date with time model used by the fundamental components.
+ */
+
+import { FdDate } from '../../calendar/models/fd-date';
 import { TimeObject } from '../../time/time-object';
 
 export class FdDatetime {
+
     date: FdDate;
     time: TimeObject;
-    
+
+    /**
+     * Static function to get the current date in FdDateTime form.
+     */
     static GetToday(): FdDatetime {
         const date: Date = new Date();
         const time: TimeObject = {hour: date.getHours(), minute: date.getMinutes(), second: date.getSeconds()};
         return new FdDatetime(FdDate.getToday(), time);
     }
 
+
+    /**
+     * Constructor to build a FdDateTime object from a FdDate and TimeObject.
+     * @param date the FdDate object.
+     * @param time the TimeObject object.
+     */
     constructor(
         date: FdDate,
         time: TimeObject
@@ -19,6 +33,9 @@ export class FdDatetime {
         this.time = time;
     }
 
+    /**
+     * Get native date object from FdDateTime.
+     */
     public toDate(): Date {
         return new Date(
             this.date.year,
@@ -30,6 +47,9 @@ export class FdDatetime {
         );
     }
 
+    /**
+     * Get native date object converted to string from FdDateTime object.
+     */
     public toLocaleDateString(): string {
         if (this.toDate()) {
             return this.toDate().toLocaleString()
@@ -38,6 +58,9 @@ export class FdDatetime {
         }
     }
 
+    /**
+     * Method that checks validity of time on FdDateTime object.
+     */
     public isTimeValid(): boolean {
         if (!this.time) {
             return false;
@@ -58,6 +81,9 @@ export class FdDatetime {
         return true;
     }
 
+    /**
+     * Method that checks validity of date on FdDateTime object.
+     */
     public isDateValid(): boolean {
         return this.date && this.date.isDateValid();
     }
